+++ pydidas/widgets/dialogues/acknowledge_box.py
@@ -39,7 +39,6 @@ from ...core.constants import (
 from ...core.utils import format_input_to_multiline_str
 from ...resources import icons
 from ..factory import CreateWidgetsMixIn
-from ..scroll_area import ScrollArea
 
 
 class AcknowledgeBox(QtWidgets.QDialog, CreateWidgetsMixIn):
@@ -51,64 +50,69 @@ class AcknowledgeBox(QtWidgets.QDialog, CreateWidgetsMixIn):
     *args : tuple
         Arguments passed to QtWidgets.QDialogue instanciation.
     **kwargs : dict
-        Keyword arguments passed to QtWidgets.QDialogue instanciation.
+        Supported keywords are:
+
+        text : str, optional
+            The notice text to be displayed. The text will be formatted to a
+            line length of a maximum of 60 characters.
+        text_preformatted : bool, optional
+            Flag to keep the text formatting and not to auto-format the text
+            to keep a width of < 60 characters.
+        title : str, optional
+            The title of the box. The default is "Notice"
     """
 
     def __init__(self, *args: tuple, **kwargs: dict):
         _text = kwargs.pop("text", "")
+        _text_preformatted = kwargs.pop("text_preformatted", False)
+        _title = kwargs.pop("title", "Notice")
+        _show_checkbox = kwargs.pop("show_checkbox", True)
         _qtapp = QtWidgets.QApplication.instance()
         QtWidgets.QDialog.__init__(self, *args, **kwargs)
         CreateWidgetsMixIn.__init__(self)
-        self.setWindowTitle("Notice")
+        self.setWindowTitle(_title)
         self.setWindowIcon(icons.pydidas_icon())
         _layout = QtWidgets.QGridLayout()
         self.setLayout(_layout)
 
         self.create_label(
             "title",
-            "Notice:",
+            f"{_title}:",
             bold=True,
             fontsize_offset=2,
-            gridPos=(0, 0, 1, 1),
+            gridPos=(0, 0, 1, 2),
         )
         self.create_label(
             "label",
             "",
             font_metric_width_factor=FONT_METRIC_WIDE_CONFIG_WIDTH,
             font_metric_height_factor=3,
+            gridPos=(1, 0, 1, 2),
             indent=8,
-            parent_widget=None,
             sizePolicy=POLICY_EXP_EXP,
             textInteractionFlags=QtCore.Qt.TextSelectableByMouse,
         )
 
-        self.create_any_widget(
-            "scroll_area",
-            ScrollArea,
-            fixedWidth=_qtapp.font_char_width * FONT_METRIC_WIDE_CONFIG_WIDTH
-            + _qtapp.scrollbar_width,
-            gridPos=(1, 0, 1, 2),
-            widget=self._widgets["label"],
-        )
         self.add_any_widget(
             "acknowledge",
             QtWidgets.QCheckBox("Do not show this notice again"),
-            gridPos=(2, 0, 1, 2),
+            gridPos=(2, 0, 1, 1),
+            visible=_show_checkbox,
         )
         self.create_button(
             "button_okay",
             "Acknowledge",
             clicked=self.close,
             font_metric_width_factor=FONT_METRIC_MEDIUM_BUTTON_WIDTH,
-            gridPos=(2, 3, 1, 1),
+            gridPos=(2, 2, 1, 1),
         )
-        self.set_text(_text)
-        self.resize(
-            _qtapp.font_char_width * (FONT_METRIC_CONSOLE_WIDTH + 10),
-            _qtapp.font_height * 10,
+        self.set_text(_text, pre_formatted=_text_preformatted)
+
+        self.setFixedWidth(
+            int(_qtapp.font_char_width * (FONT_METRIC_CONSOLE_WIDTH + 10))
         )
 
-    def set_text(self, text: str):
+    def set_text(self, text: str, pre_formatted: bool = False):
         """
         Set the text in the message box.
 
@@ -116,9 +120,13 @@ class AcknowledgeBox(QtWidgets.QDialog, CreateWidgetsMixIn):
         ----------
         text : str
             The text to be displayed.
+        pre_formatted : bool, optional
+            Flag to keep the text formatting. The default is False.
         """
-        _new_text = format_input_to_multiline_str(text, max_line_length=60)
-        self._widgets["label"].setText(_new_text)
+        if not pre_formatted:
+            text = format_input_to_multiline_str(text, max_line_length=60)
+        self._widgets["label"].setText(text)
+        self._widgets["label"].font_metric_height_factor = len(text.split("\n"))
 
     @QtCore.Slot()
     def close(self):